Add config file syntaxes
This commit adds support for highlighting header files used in
conjunction with CMake's "configure_file" command. If you have a file
that ends with *.h.in, then that will be highlighted with the CMake C
Header syntax. That syntax is simply the C syntax, but with CMake
variable substition syntax and the #cmakedefine preprocessor directive
highlighted. The same holds for C++ header files that end with *.in.