Add all builtin commands together with their keywords
The way this works is that we maintain a YML file with the builtin commands,
and then use tools/update-commands to generate the
CMakeCommands.sublime-syntax file. This file is included in the
CMake.sublime-syntax file.

Also, some scopes are changed: in practice having almost everything as a
string.unquoted was annoying, so I changed it into meta.string.unquoted
to signify that some things are unquoted strings, but they don't get a
color in this way.

The set command now assigns variable.other.readwrite.assignment to its first
argument for easy lookup.